+++ crates/rslint_errors/src/codespan/term/views.rs
@@ -397,7 +397,7 @@ where
         //      found type `String`
         // ```
         for note in &self.diagnostic.notes {
-            renderer.render_snippet_note(outer_padding, note)?;
+            renderer.render_snippet_note(outer_padding, note.to_owned())?;
         }
         Ok(())
     }
@@ -474,7 +474,7 @@ where
             //      found type `String`
             // ```
             for note in &self.diagnostic.notes {
-                renderer.render_snippet_note(0, note)?;
+                renderer.render_snippet_note(0, note.to_owned())?;
             }
         }