+++ crates/rslint_errors/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed 
+
+- Changed codespan backend to render notes with severity correctly
+
 ## [0.1.1]
 
 ### Added